Skip running `20260515000001_add_store_id_to_spree_newsletter_subscribers.rb` on spree_multi_tenant

+++ spree/core/db/migrate/20260515000001_add_store_id_to_spree_newsletter_subscribers.rb
@@ -2,15 +2,11 @@ class AddStoreIdToSpreeNewsletterSubscribers < ActiveRecord::Migration[7.2]
   def up
     add_reference :spree_newsletter_subscribers, :store
 
-    if defined?(SpreeMultiTenant)
-      Spree::Tenant.find_each do |tenant|
-        SpreeMultiTenant.with_tenant(tenant) do
-          Spree::NewsletterSubscriber.where(store_id: nil).update_all(store_id: Spree::Store.default.id)
-        end
-      end
-    else
-      Spree::NewsletterSubscriber.update_all(store_id: Spree::Store.default.id)
-    end
+    # For spree_multi_tenant we need handle the backfill and indices there
+    return if defined?(SpreeMultiTenant)
+
+    default_store = Spree::Store.default
+    Spree::NewsletterSubscriber.update_all(store_id: default_store.id) if default_store&.persisted?
 
     change_column_null :spree_newsletter_subscribers, :store_id, false
 
@@ -19,8 +15,11 @@ class AddStoreIdToSpreeNewsletterSubscribers < ActiveRecord::Migration[7.2]
   end
 
   def down
-    remove_index :spree_newsletter_subscribers, [:email, :store_id], if_exists: true
-    add_index :spree_newsletter_subscribers, :email, unique: true, if_not_exists: true
+    unless defined?(SpreeMultiTenant)
+      remove_index :spree_newsletter_subscribers, [:email, :store_id], if_exists: true
+      add_index :spree_newsletter_subscribers, :email, unique: true, if_not_exists: true
+    end
+
     remove_reference :spree_newsletter_subscribers, :store
   end
 end
